backend/fastapi/ai/ocr/requirements.txt
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
bokeh==3.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
finance-datareader==0.9.96
fonttools==4.56.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
OpenDartReader==0.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tifffile==2025.3.13
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2025.1.0
